Projects/AI-Flashcard/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.26.0
beautifulsoup4>=4.10.0
nltk>=3.6.7
python-dotenv>=0.19.0
openai>=1.0.0
youtube-transcript-api>=0.6.1
lxml>=4.9.2
ollama